requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aioredis==2.0.1
aiosignal==1.3.2
aiosmtplib==3.0.2
aiosqlite==0.20.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
Authlib==1.3.2
babel==2.16.0
bcrypt==4.2.1
blinker==1.9.0
boto3==1.38.2
botocore==1.38.2
certifi==2024.8.30
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
Deprecated==1.2.18
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-mail==1.4.2
fastapi-storages==0.3.0
frozenlist==1.7.0
gigachat==0.1.40
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
limits==5.4.0
magic-filter==1.0.12
Mako==1.3.8
MarkupSafe==3.0.2
multidict==6.6.3
numpy==2.2.1
packaging==25.0
pandas==2.2.3
passlib==1.7.4
propcache==0.3.2
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
PyJWT==2.10.1
pypika-tortoise==0.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.19
pytz==2024.2
PyYAML==6.0.2
rsa==4.9
s3transfer==0.12.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
sqladmin==0.20.1
SQLAlchemy==2.0.40
starlette==0.46.2
time-machine==2.16.0
tortoise-orm==0.22.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2024.2
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.1.2
yarl==1.20.1
